+++ nuget/OpenTelemetry.AutoInstrumentation/contentFiles/any/any/instrument.sh
@@ -11,6 +11,5 @@ export DOTNET_STARTUP_HOOKS=${BASE_PATH}/OpenTelemetry.AutoInstrumentation.Start
 
 # Settings for OpenTelemetry
 export OTEL_DOTNET_AUTO_HOME=${BASE_PATH}
-export OTEL_DOTNET_AUTO_INTEGRATIONS_FILE=${BASE_PATH}/integrations.json
 
 $@