+++ src/OpenTelemetry.AutoInstrumentation.Native/rejit_handler.h
@@ -8,16 +8,24 @@
 
 #include <atomic>
 #include <mutex>
+#include <shared_mutex>
 #include <string>
 #include <unordered_map>
 #include <vector>
+#include <future>
 
 #include "cor.h"
 #include "corprof.h"
 #include "module_metadata.h"
+#include "rejit_work_offloader.h"
+#include "method_rewriter.h"
 
 namespace trace
 {
+
+typedef std::shared_mutex Lock;
+typedef std::unique_lock<Lock> WriteLock;
+typedef std::shared_lock<Lock> ReadLock;
 
 // forward declarations...
 class RejitHandlerModule;
@@ -28,11 +36,10 @@ class RejitHandler;
 /// </summary>
 class RejitHandlerModuleMethod
 {
-private:
+protected:
     mdMethodDef m_methodDef;
     ICorProfilerFunctionControl* m_pFunctionControl;
     std::unique_ptr<FunctionInfo> m_functionInfo;
-    std::unique_ptr<MethodReplacement> m_methodReplacement;
 
     std::mutex m_ngenModulesLock;
     std::unordered_map<ModuleID, bool> m_ngenModules;
@@ -40,7 +47,7 @@ private:
     RejitHandlerModule* m_module;
 
 public:
-    RejitHandlerModuleMethod(mdMethodDef methodDef, RejitHandlerModule* module);
+    RejitHandlerModuleMethod(mdMethodDef methodDef, RejitHandlerModule* module, const FunctionInfo& functionInfo);
     mdMethodDef GetMethodDef();
     RejitHandlerModule* GetModule();
 
@@ -50,11 +57,32 @@ public:
     FunctionInfo* GetFunctionInfo();
     void SetFunctionInfo(const FunctionInfo& functionInfo);
 
-    MethodReplacement* GetMethodReplacement();
-    void SetMethodReplacement(const MethodReplacement& methodReplacement);
-
     void RequestRejitForInlinersInModule(ModuleID moduleId);
+    virtual MethodRewriter* GetMethodRewriter() = 0;
+
+    virtual ~RejitHandlerModuleMethod() = default;
 };
+
+/// <summary>
+/// Rejit handler representation of a method
+/// </summary>
+class TracerRejitHandlerModuleMethod : public RejitHandlerModuleMethod
+{
+private:
+    std::unique_ptr<IntegrationDefinition> m_integrationDefinition;
+
+public:
+    TracerRejitHandlerModuleMethod(
+                    mdMethodDef methodDef,
+                    RejitHandlerModule* module,
+                    const FunctionInfo& functionInfo,
+                    const IntegrationDefinition& integrationDefinition);
+    
+    IntegrationDefinition* GetIntegrationDefinition();
+    MethodRewriter* GetMethodRewriter() override;
+};
+
+using RejitHandlerModuleMethodCreatorFunc = std::function<std::unique_ptr<RejitHandlerModuleMethod>(const mdMethodDef, RejitHandlerModule*)>;
 
 /// <summary>
 /// Rejit handler representation of a module
@@ -63,7 +91,7 @@ class RejitHandlerModule
 {
 private:
     ModuleID m_moduleId;
-    ModuleMetadata* m_metadata;
+    std::unique_ptr<ModuleMetadata> m_metadata;
     std::mutex m_methods_lock;
     std::unordered_map<mdMethodDef, std::unique_ptr<RejitHandlerModuleMethod>> m_methods;
     RejitHandler* m_handler;
@@ -76,9 +104,9 @@ public:
     ModuleMetadata* GetModuleMetadata();
     void SetModuleMetadata(ModuleMetadata* metadata);
 
-    RejitHandlerModuleMethod* GetOrAddMethod(mdMethodDef methodDef);
-    bool TryGetMethod(mdMethodDef methodDef, RejitHandlerModuleMethod** methodHandler);
+    bool CreateMethodIfNotExists(const mdMethodDef methodDef, RejitHandlerModuleMethodCreatorFunc creator);
     bool ContainsMethod(mdMethodDef methodDef);
+    bool TryGetMethod(mdMethodDef methodDef, /* OUT */ RejitHandlerModuleMethod** methodHandler);
 
     void RequestRejitForInlinersInModule(ModuleID moduleId);
 };
@@ -90,39 +118,54 @@ public:
 class RejitHandler
 {
 private:
+    std::atomic_bool m_shutdown = {false};
+    Lock m_shutdown_lock;
+
     std::mutex m_modules_lock;
     std::unordered_map<ModuleID, std::unique_ptr<RejitHandlerModule>> m_modules;
+    AssemblyProperty* m_pCorAssemblyProperty = nullptr;
 
-    ICorProfilerInfo7* m_profilerInfo7;
+    ICorProfilerInfo7* m_profilerInfo;
+    ICorProfilerInfo10* m_profilerInfo10;
 
-    std::function<HRESULT(RejitHandlerModule*, RejitHandlerModuleMethod*)> m_rewriteCallback;
+    std::shared_ptr<RejitWorkOffloader> m_work_offloader;
+        
+    bool enable_by_ref_instrumentation = false;
+    bool enable_calltarget_state_by_ref = false;
 
     std::mutex m_ngenModules_lock;
     std::vector<ModuleID> m_ngenModules;
 
     void RequestRejitForInlinersInModule(ModuleID moduleId);
-
 public:
-    RejitHandler(ICorProfilerInfo7* pInfo,
-                 std::function<HRESULT(RejitHandlerModule*, RejitHandlerModuleMethod*)> rewriteCallback);
+    RejitHandler(ICorProfilerInfo7* pInfo, std::shared_ptr<RejitWorkOffloader> work_offloader);
+    RejitHandler(ICorProfilerInfo10* pInfo, std::shared_ptr<RejitWorkOffloader> work_offloader);
 
     RejitHandlerModule* GetOrAddModule(ModuleID moduleId);
+    void SetEnableByRefInstrumentation(bool enableByRefInstrumentation);
+    void SetEnableCallTargetStateByRef(bool enableCallTargetStateByRef);
+    bool GetEnableCallTargetStateByRef();
+    bool GetEnableByRefInstrumentation();
 
-    bool TryGetModule(ModuleID moduleId, RejitHandlerModule** moduleHandler);
     void RemoveModule(ModuleID moduleId);
+    bool HasModuleAndMethod(ModuleID moduleId, mdMethodDef methodDef);
 
     void AddNGenModule(ModuleID moduleId);
 
+    void EnqueueForRejit(std::vector<ModuleID>& modulesVector, std::vector<mdMethodDef>& modulesMethodDef);
     void RequestRejit(std::vector<ModuleID>& modulesVector, std::vector<mdMethodDef>& modulesMethodDef);
 
     void Shutdown();
+    bool IsShutdownRequested();
 
     HRESULT NotifyReJITParameters(ModuleID moduleId, mdMethodDef methodId,
-                                  ICorProfilerFunctionControl* pFunctionControl, ModuleMetadata* metadata);
+                                  ICorProfilerFunctionControl* pFunctionControl);
     HRESULT NotifyReJITCompilationStarted(FunctionID functionId, ReJITID rejitId);
 
-    ICorProfilerInfo7* GetCorProfilerInfo7();
+    ICorProfilerInfo7* GetCorProfilerInfo();
 
+    void SetCorAssemblyProfiler(AssemblyProperty* pCorAssemblyProfiler);
+    AssemblyProperty* GetCorAssemblyProperty();
     void RequestRejitForNGenInliners();
 };