+++ src/OpenTelemetry.AutoInstrumentation/CallTarget/Handlers/Continuations/ValueTaskContinuationGenerator`1.cs
@@ -20,7 +20,7 @@ namespace OpenTelemetry.AutoInstrumentation.CallTarget.Handlers.Continuations;
 #if NET6_0_OR_GREATER
 internal class ValueTaskContinuationGenerator<TIntegration, TTarget, TReturn, TResult> : ContinuationGenerator<TTarget, TReturn>
 {
-    private static readonly Func<TTarget, TResult?, Exception?, CallTargetState, TResult>? _continuation;
+    private static readonly ContinuationMethodDelegate? _continuation;
     private static readonly bool _preserveContext;
 
     static ValueTaskContinuationGenerator()
@@ -28,12 +28,14 @@ internal class ValueTaskContinuationGenerator<TIntegration, TTarget, TReturn, TR
         var result = IntegrationMapper.CreateAsyncEndMethodDelegate(typeof(TIntegration), typeof(TTarget), typeof(TResult));
         if (result.Method != null)
         {
-            _continuation = (Func<TTarget, TResult?, Exception?, CallTargetState, TResult>)result.Method.CreateDelegate(typeof(Func<TTarget, TResult?, Exception?, CallTargetState, TResult>));
+            _continuation = (ContinuationMethodDelegate)result.Method.CreateDelegate(typeof(ContinuationMethodDelegate));
             _preserveContext = result.PreserveContext;
         }
     }
 
-    public override TReturn? SetContinuation(TTarget instance, TReturn? returnValue, Exception? exception, CallTargetState state)
+    internal delegate TResult ContinuationMethodDelegate(TTarget target, TResult? returnValue, Exception? exception, in CallTargetState state);
+
+    public override TReturn? SetContinuation(TTarget instance, TReturn? returnValue, Exception? exception, in CallTargetState state)
     {
         if (_continuation is null)
         {
@@ -42,7 +44,7 @@ internal class ValueTaskContinuationGenerator<TIntegration, TTarget, TReturn, TR
 
         if (exception != null)
         {
-            _continuation(instance, default, exception, state);
+            _continuation(instance, default, exception, in state);
             return returnValue;
         }
 
@@ -63,7 +65,7 @@ internal class ValueTaskContinuationGenerator<TIntegration, TTarget, TReturn, TR
                     // *
                     // Calls the CallTarget integration continuation, exceptions here should never bubble up to the application
                     // *
-                    _continuation!(instance, result, ex, state);
+                    _continuation!(instance, result, ex, in state);
                 }
                 catch (Exception contEx)
                 {
@@ -78,7 +80,7 @@ internal class ValueTaskContinuationGenerator<TIntegration, TTarget, TReturn, TR
                 // *
                 // Calls the CallTarget integration continuation, exceptions here should never bubble up to the application
                 // *
-                return _continuation!(instance, result, null, state);
+                return _continuation!(instance, result, null, in state);
             }
             catch (Exception contEx)
             {