+++ src/OpenTelemetry.AutoInstrumentation/NativeMethods.cs
@@ -32,12 +32,82 @@ internal static class NativeMethods
         return NonWindows.IsProfilerAttached();
     }
 
+    public static void InitializeProfiler(string id, NativeCallTargetDefinition[] methodArrays)
+    {
+        if (methodArrays is null || methodArrays.Length == 0)
+        {
+            return;
+        }
+
+        if (IsWindows)
+        {
+            Windows.InitializeProfiler(id, methodArrays, methodArrays.Length);
+        }
+        else
+        {
+            NonWindows.InitializeProfiler(id, methodArrays, methodArrays.Length);
+        }
+    }
+
+    public static void EnableByRefInstrumentation()
+    {
+        if (IsWindows)
+        {
+            Windows.EnableByRefInstrumentation();
+        }
+        else
+        {
+            NonWindows.EnableByRefInstrumentation();
+        }
+    }
+
+    public static void EnableCallTargetStateByRef()
+    {
+        if (IsWindows)
+        {
+            Windows.EnableCallTargetStateByRef();
+        }
+        else
+        {
+            NonWindows.EnableCallTargetStateByRef();
+        }
+    }
+
+    public static void AddDerivedInstrumentations(string id, NativeCallTargetDefinition[] methodArrays)
+    {
+        if (methodArrays is null || methodArrays.Length == 0)
+        {
+            return;
+        }
+
+        if (IsWindows)
+        {
+            Windows.AddDerivedInstrumentations(id, methodArrays, methodArrays.Length);
+        }
+        else
+        {
+            NonWindows.AddDerivedInstrumentations(id, methodArrays, methodArrays.Length);
+        }
+    }
+
     // the "dll" extension is required on .NET Framework
     // and optional on .NET Core
     private static class Windows
     {
         [DllImport("OpenTelemetry.AutoInstrumentation.Native.dll")]
         public static extern bool IsProfilerAttached();
+
+        [DllImport("OpenTelemetry.AutoInstrumentation.Native.dll")]
+        public static extern void InitializeProfiler([MarshalAs(UnmanagedType.LPWStr)] string id, [In] NativeCallTargetDefinition[] methodArrays, int size);
+
+        [DllImport("OpenTelemetry.AutoInstrumentation.Native.dll")]
+        public static extern void EnableByRefInstrumentation();
+
+        [DllImport("OpenTelemetry.AutoInstrumentation.Native.dll")]
+        public static extern void EnableCallTargetStateByRef();
+
+        [DllImport("OpenTelemetry.AutoInstrumentation.Native.dll")]
+        public static extern void AddDerivedInstrumentations([MarshalAs(UnmanagedType.LPWStr)] string id, [In] NativeCallTargetDefinition[] methodArrays, int size);
     }
 
     // assume .NET Core if not running on Windows
@@ -45,5 +115,17 @@ internal static class NativeMethods
     {
         [DllImport("OpenTelemetry.AutoInstrumentation.Native")]
         public static extern bool IsProfilerAttached();
+
+        [DllImport("OpenTelemetry.AutoInstrumentation.Native")]
+        public static extern void InitializeProfiler([MarshalAs(UnmanagedType.LPWStr)] string id, [In] NativeCallTargetDefinition[] methodArrays, int size);
+
+        [DllImport("OpenTelemetry.AutoInstrumentation.Native")]
+        public static extern void EnableByRefInstrumentation();
+
+        [DllImport("OpenTelemetry.AutoInstrumentation.Native")]
+        public static extern void EnableCallTargetStateByRef();
+
+        [DllImport("OpenTelemetry.AutoInstrumentation.Native")]
+        public static extern void AddDerivedInstrumentations([MarshalAs(UnmanagedType.LPWStr)] string id, [In] NativeCallTargetDefinition[] methodArrays, int size);
     }
 }